+++ app/components/FormInput.jsx
@@ -1,0 +1,11 @@
+export default function FormInput({ type, value, onChange, placeholder }) {
+  return (
+    <input
+      type={type}
+      value={value}
+      onChange={onChange}
+      placeholder={placeholder}
+      className="w-full px-4 py-2 border border-gray-300 rounded focus:outline-none focus:ring-2 focus:ring-blue-500 text-black"
+    />
+  );
+}
