+++ app/page.jsx
@@ -1,112 +1,47 @@
 "use client";
 
-import { useState, useEffect } from "react";
-import { useRouter } from "next/navigation";
+import ErrorText from "./components/ErrorText";
+import useLoginForm from "./hooks/useLoginForm";
+
+import FormInput from "./components/FormInput";
 
 export default function LoginPage() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [emailError, setEmailError] = useState("");
-  const [passwordError, setPasswordError] = useState("");
-  const [serverError, setServerError] = useState("");
-  const router = useRouter();
-
-  //토큰 있을 시 /service 페이지로 바로이동
-  useEffect(() => {
-    const token = localStorage.getItem("token");
-    if (token) {
-      window.location.href = "/service";
-    }
-  }, []);
-//유효성검사
-  const validateForm = () => {
-    let valid = true;
-
-    if (!email.includes("@")) {
-      setEmailError("유효한 이메일 주소를 입력하세요.");
-      valid = false;
-    } else {
-      setEmailError("");
-    }
-
-    const passwordRegex = /^\d{8,}$/;
-    if (!passwordRegex.test(password)) {
-      setPasswordError("비밀번호는 숫자만 포함하며 최소 8자리여야 합니다.");
-      valid = false;
-    } else {
-      setPasswordError("");
-    }
-
-    return valid;
-  };
-  //로그인 정보 전송
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (!validateForm()) return;
-
-    try {
-      const res = await fetch("/auth/", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ email, password }),
-      });
-
-      if (res.ok) {
-        const { token, serviceUrl } = await res.json();
-        localStorage.setItem("token", token);
-        window.location.href = serviceUrl;
-      } else {
-        const data = await res.json();
-        setServerError(data.error || "로그인 실패");
-      }
-    } catch (err) {
-      setServerError("서버와 통신 중 오류가 발생했습니다.");
-    }
-  };
+  const {
+    email, setEmail,
+    password, setPassword,
+    emailError, passwordError, serverError,
+    handleSubmit
+  } = useLoginForm();
 
   return (
     <div className="min-h-screen flex items-center justify-center bg-gray-100 px-4">
       <div className="w-full max-w-sm p-8 bg-white shadow-lg rounded-lg">
-        <h1 className="text-2xl font-semibold text-center text-gray-800 mb-6">
-          로그인
-        </h1>
+        <h1 className="text-2xl font-semibold text-center text-gray-800 mb-6">로그인</h1>
         <form onSubmit={handleSubmit} className="space-y-4">
           <div>
-            <input
+            <FormInput
               type="email"
               value={email}
               onChange={(e) => setEmail(e.target.value)}
               placeholder="이메일"
-              className="w-full px-4 py-2 border border-gray-300 rounded focus:outline-none focus:ring-2 focus:ring-blue-500 text-black"
             />
-            {emailError && (
-              <p className="text-red-500 text-sm mt-1">{emailError}</p>
-            )}
+            <ErrorText message={emailError} />
           </div>
 
           <div>
-            <input
+            <FormInput
               type="password"
               value={password}
               onChange={(e) => setPassword(e.target.value)}
               placeholder="비밀번호"
-              className="w-full px-4 py-2 border border-gray-300 rounded focus:outline-none focus:ring-2 focus:ring-blue-500 text-black"
             />
-            {passwordError && (
-              <p className="text-red-500 text-sm mt-1">{passwordError}</p>
-            )}
+            <ErrorText message={passwordError} />
           </div>
 
-          <button
-            type="submit"
-            className="w-full py-2 bg-blue-600 text-white rounded hover:bg-blue-700 transition"
-          >
+          <button type="submit" className="w-full py-2 bg-blue-600 text-white rounded hover:bg-blue-700 transition">
             로그인
           </button>
-          {serverError && (
-            <p className="text-red-500 text-sm text-center mt-2">{serverError}</p>
-          )}
+          <ErrorText message={serverError} center />
         </form>
       </div>
     </div>
