+++ app/utils/decodeJWT.js
@@ -1,0 +1,7 @@
+export function decodeJWT(token) {
+  try {
+    return JSON.parse(atob(token.split(".")[1]));
+  } catch {
+    return null;
+  }
+}
